models/requirements.txt
langchain
langchain-core
langchain-google-genai
google-generativeai
langchain-huggingface
huggingface-hub
python-dotenv
numpy
scikit-learn